requirements.txt
streamlit 
st-annotated-text
numpy
pandas
sacremoses
sentencepiece
tokenizers
torch
transformers
spacy==2.3.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz